Added some validation in friends league join and delete

+++ controller/friendsLeagueController.js
@@ -46,14 +46,29 @@ module.exports.createFriendsLeague = async (request, response) => {
         response.send("Friends League not created");
     }
 };
-
+async function checkIsAutoJoin(id){
+    const fl = await friendsLeagueDatabase.getFriendsLeagueByID(id);
+    if(fl == null){
+        return false;
+    }
+    return fl.auto_join;
+}
 module.exports.joinFriendsLeague = async (request, response) => {
     try{
-        const ret = await friendsLeagueDatabase.joinFriendsLeague(request.user.user_id, request.body.id);
-        if(ret){
-            response.send("Joined Friends League");
+        if(!(await checkIsAutoJoin(request.body.id))){
+            const ret = await friendsLeagueDatabase.requestFriendsLeague(request.user.user_id, request.body.id);
+            if(ret){
+                response.send("Requested to join Friends League");
+            }else{
+                response.send("Not Requested to join Friends League");
+            }
         }else{
-            response.send("Not Joined Friends League");
+            const ret = await friendsLeagueDatabase.joinFriendsLeague(request.user.user_id, request.body.id);
+            if(ret){
+                response.send("Joined Friends League");
+            }else{
+                response.send("Not Joined Friends League");
+            }
         }
     }catch(error){
         response.send("Not Joined Friends League");
@@ -72,9 +87,26 @@ module.exports.leaveFriendsLeague = async (request, response) => {
         response.send("Not Left Friends League");
     }
 }
+async function checkIsAdmin(user_id, fl_id){
+    const fl = await friendsLeagueDatabase.getFriendsLeagueByID(fl_id);
+    if(fl == null){
+        return false;
+    }
+    const allMember = await friendsLeagueDatabase.getFriendsLeagueTeams(request.body.id);
+    for(var i = 0; i < allMember.length; i++){
+        if(allMember[i].user_id == request.user.user_id && allMember[i].role == 'admin'){
+            return true
+        }
+    }
+    return false;
 
+}
 module.exports.deleteFriendsLeague = async (request, response) => {
     try{
+        if(!await checkIsAdmin(request.user.user_id, request.body.id)){
+            response.send("Not Deleted Friends League");
+            return;
+        }
         const ret = await friendsLeagueDatabase.deleteFriendsLeague(request.body.id);
         if(ret){
             response.send("Deleted Friends League");
